refactor: polish autogen.beta imports
