+++ autogen/beta/__init__.py
@@ -23,7 +23,6 @@ from .spec import AgentSpec
 from .stream import MemoryStream
 from .task import Task, TaskInject, TaskSpec
 from .tools import ToolResult, Toolkit, tool
-from .usage import UsageRecord, UsageReport
 
 __all__ = (
     "Agent",
@@ -50,8 +49,6 @@ __all__ = (
     "TextInput",
     "ToolResult",
     "Toolkit",
-    "UsageRecord",
-    "UsageReport",
     "Variable",
     "VideoInput",
     "observer",
